refactor : event 하나의 파일로 이동
